Add User tests
