+++ dart_test.yaml
@@ -43,3 +43,5 @@ tags:
   depends_on_RssFeedEpisode:
   depends_on_RssFeedMetadata:
   depends_on_SeriesProgress:
+  depends_on_UserPermissions:
+  depends_on_PlaybackSession:
